+++ src/AmmoFinder.RateLimiter/RateLimiterEngine.cs
@@ -1,9 +1,7 @@
 ﻿using Polly;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
-using System.Threading.Tasks;
 
 namespace AmmoFinder.RateLimiter
 {
@@ -17,45 +15,11 @@ namespace AmmoFinder.RateLimiter
             int limitCount,
             TimeSpan limitTime)
         {
-            var now = DateTimeOffset.UtcNow;
-
-            lock (callLog)
-            {
-                callLog.Add(now);
-
-                while (callLog.Count > limitCount)
-                    callLog.RemoveAt(0);
-            }
-
-            LimitDelay(now, callLog, limitCount, limitTime).RunSynchronously();
+            Delayer.LimitDelay(callLog, limitCount, limitTime).RunSynchronously();
 
             TResult result = action(context, cancellationToken);
 
             return result;
-        }
-
-        private static async Task LimitDelay(DateTimeOffset now, List<DateTimeOffset> callLog, int limitCount, TimeSpan limitTime)
-        {
-            if (callLog.Count < limitCount)
-                return;
-
-            var limit = now.Add(-limitTime);
-
-            var lastCall = DateTimeOffset.MinValue;
-            var shouldLock = false;
-
-            lock (callLog)
-            {
-                lastCall = callLog.FirstOrDefault();
-                shouldLock = callLog.Count(x => x >= limit) >= limitCount;
-            }
-
-            var delayTime = shouldLock && (lastCall > DateTimeOffset.MinValue)
-                ? (limit - lastCall)
-                : TimeSpan.Zero;
-
-            if (delayTime > TimeSpan.Zero)
-                await Task.Delay(delayTime);
         }
     }
 }
